Add seed data, be independent from Drupal

+++ packages/eci-cli/src/commands/api/clean.js
@@ -3,12 +3,23 @@ module.exports = {
   run: async toolbox => {
     const {
       print: { info, error },
+      parameters: { options },
       system: { run },
     } = toolbox;
 
-    info('Cleaning previously saved data from Drupal JSON:API ...');
+    let folder =
+      options.folder ||
+      process.env.DRUPAL_JSONAPI_OFFLINE_FOLDER ||
+      'api/drupal/jsonapi';
 
-    const dir = `${__dirname}/drupal_jsonapi`;
+    const dir = folder.split('/')[0]; // only first segment is needed
+
+    if (!dir) {
+      info("Missing folder to delete. Can't do anything.");
+      return;
+    }
+
+    info('Cleaning previously saved data from Drupal JSON:API ...');
 
     try {
       await run(`rm -rf ${dir}`);
